Add files via upload
Hope you will enjoy this game. All the guidelines are given in the first sheet(Intro) follow the guidelines and enjoy the game. The main thing is Read the story and then answer the question that will give you a real Treasure Hunt experience else it will be like a Quiz game. All the best. Comment how much time you took to solve this Mystery.